fix: use size_t in escapeSpecialChars() to prevent int overflow
Using int for string lengths allows overflow in len * 2 + 1 for large
inputs (INT_MAX/2 byte strings). Changed loop variables and len to size_t
to match malloc and strlen return types.

+++ src/bialet_wren.c
@@ -395,8 +395,8 @@ static void queryExecute(WrenVM* vm, BialetQuery* query) {
   sqlite3_finalize(stmt);
 }
 char* escapeSpecialChars(const char* input) {
-  int   i, j = 0, len = strlen(input);
-  char* output = malloc(len * 2 + 1); // Worst case: all characters need escaping
+  size_t i, j = 0, len = strlen(input);
+  char* output = malloc(len * 2 + 1);
   if(output == NULL)
     return NULL;
